+++ ProgressBar/ProgressBar.js
@@ -139,7 +139,7 @@ const ProgressBarBase = kind({
 
 	styles: {
 		css: componentCss,
-		publicClassNames: ['progressBar', 'radial', 'bar', 'fill']
+		publicClassNames: ['progressBar', 'radial', 'bar', 'fill', 'seperator']
 	},
 
 	computed: {